move AssetKeyOrCheckKey and AssetCheckKey to asset_key.py
